Update patch set 1
Patch Set 1: Code-Review-2

The point of the watchdog timer is to ensure that everything else on the system is NOT hung or hogging resources.

With this change, any other task doing a while(1); will not trigger the watchdog.  This defeats the purpose of the watchdog.

If 'b 100' is causing a watchdog reboot, we need to fix the underlying cause, not treat the symptom.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2