Falco: correct input current limit for 45W adapter, Turbo off
Change the charger's Input Current Register setting for the 45W adapter to
match the latest spec.

BUG=chrome-os-partner:20739
BRANCH=Falco,ToT
TEST=manual

Connect a 45W adapter, run the "battery" and "charger" commands on the EC
console.

When the battery charge is below 10% (turbo off), the "I_in" value displayed
by the "charger" command should be 1536. Before it was 2560.

Change-Id: I0483b5408aa2da352cd3aeda58e1656c095d86b2
Signed-off-by: Bill Richardson <wfrichar@chromium.org>