spring: increase the initial current limit
Put a higher initial current limit to avoid browning out the system
when we turn on charging, lower power bricks might cut off but
we will re-enable them with a lower current limit later.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:22807
TEST=On Spring, reset the EC with various chargers plugged,
and see that the charging subsystem is stable in all cases even though
the brick might cut the current initially.

Change-Id: I2bf7bace258a69a5c11b4475b1447d6d2e5e53a7