Battery command needs to delay between iterations
The battery command takes a long time to run if given a high repeat
count.  Since it doesn't sleep at all between iterations by default,
this will unsurprisingly starve all other tasks and cause a watchdog
timeout.

Reset the watchdog between iterations, to give the rest of the system time
to do things.  This is similar to what we do in the i2cscan command.

BUG=chrome-os-partner:22232
BRANCH=none
TEST=apshutdown, then battery 1000 = no watchdog

Change-Id: I3ce55e15d90a6dfda34b1e2e332d7f7828922e78
Signed-off-by: Randall Spangler <rspangler@chromium.org>