spring: allow to read battery specific registers from AP
Extends the range of battery registers which can be read by the AP with
the passthrough, from the subset of registers defined in the smart
battery standard to the full range of vendor specific registers for the
Spring battery.

The change allows to read the individual cell voltages from the AP if we
want to.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:22807
TEST=none

Change-Id: Ifca0d137484bdc5629515f9b67a396cb6a5f35c5