Clean up G781 temperature sensor power detection
This was previously done in a board-specific function across 4 boards.
Except that the board-specific function was identical in all cases
(that is, not really board-specific).  Put it back in the common
implementation to get rid of duplicated code.

If the sensor is ever on a different-named rail on a board, we'll add
a CONFIG_TEMP_SENSOR_G781_POWER_GPIO or
CONFIG_TEMP_SENSOR_G781_POWER_BOARD_SPECIFIC config option which does
something board-specific only for the board(s) which need it.

BUG=chrome-os-partner:18343
BRANCH=none
TEST=build all boards; pass unit tests

Change-Id: I29de40001d5d4dc873e5ba8f3abb328c6271f235
Signed-off-by: Randall Spangler <rspangler@chromium.org>