Add config for USB power rail GPIO
Some systems have a separate GPIO for the USB power rail. If so, turn that
rail off when the AP shuts down, and turn it back on when it comes up again.

Note that this is a board-specific config for power sequencing, not a
generic run-time power conservation feature.

BUG=chrome-os-partner:22870
BRANCH=none
TEST=manual

  make runtests

Nothing actually uses this yet, so there's nothing to test.

Change-Id: I51c00b2ca56c57961bcbd332be1e692bcb2745ae
Signed-off-by: Bill Richardson <wfrichar@chromium.org>