Update patch set 1
Patch Set 1: Code-Review-1

This works only in the case where ports are turned all-off or all-on.

If the host turns off port 1, then turns off port 2, all ports are off but the GPIO is still on.

Seems like we need to set the GPIO to (any port on ? 1 : 0)

Patch-set: 1
Label: Code-Review=-1