Update patch set 1
Patch Set 1:

Yeah, but the RESUME and SHUTDOWN hooks are what turn the GPIO on and off. So the USB power goes on when the system comes up, and goes off again when it shuts down (S5).

We're not (yet) controlling the power based on whether or not any of the ports need it. That can't happen unless the AP is in S0, anyway.

Patch-set: 1