spring: Prevent cutting off power on sysjump
If the battery is dead, we must ensure the input current is not cut off
on sysjump. To do so, we record the time of the jump and the I_LIM PWM
duty cycle when jumping. Within 20 seconds from the jump, restrict the
input current to be equal to or higher than that when jumping. This will
ensure the system have at least the same amount of power as when sysjump
happens.

If the user switches to a low power charger within the 20-second
interval, the worst case is that the charger browns out and then recover
after 20 seconds, which should be fine.

BUG=chrome-os-partner:23055
TEST=Unplug battery and do 'sysjump RW' in S0. Without this CL, the AP
shuts down and the EC browns out. With this CL, the AP remains in S0.
BRANCH=Spring

Change-Id: Icf12cd2ec3a59786c01e0b1464a6f094e7c12756
Signed-off-by: Vic Yang <victoryang@chromium.org>