spring: do not start CPU when battery is self cut-off
When the battery self cuts off due to a cell under-voltage, the battery
discharging FET is disabled but the gas gauge still advertise available
capacity.
The patch specifically detects that case and prevents the main CPU from
running since the battery is doing any power buffering until the DSG FET
is enabled.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:22807
TEST=On spring, with two good battery (one somewhat charged, the other
discharged) successfully starts the machine plugged on Spring charger.
With a self cut-off battery, see the EC preventing the CPU start-up and
the system staying stable and charging the battery.

Change-Id: I1eec1a6f061bbf3b6e4ca996cecbe1ad76c0f29d