falco: Adjust low-battery LED blink threshold (again)
Something changed somewhere, and now we're off by a few points again. The
battery LED is supposed to start blinking when the charge is under 12%, but
that level is pretty fuzzy. For example:

I force the battery level to 11% using the EC console:

  > battfake 11
  Reporting fake battery level 11%
  >

The host's power_supply_info command shows:

  percentage:         10.9884
  display percentage: 11.3282

but clicking on the battery icon on the GUI says

  9% - 0-:48 left

So we're off by two percentage points. However, at 20%:

  > battfake 20
  Reporting fake battery level 20%
  >

I get

  percentage:         19.9943
  display percentage: 20.6127

and

  19% - 1:30 left

which is much closer.

It's still nonlinear, though.

If I force the battery to 15%, the GUI reports 14%.
If I force the battery to 14%, the GUI reports 11%.

Whatever. We'll put the EC's threshold at 14%, so that the light starts
blinking when the GUI is below 12%. Note that the GUI only updates about
once a minute or so, so you'll have to be patient.

BUG=chrome-os-partner:22159
BRANCH=falco
TEST=manual

Let the battery discharge, watch for the light to start flashing when the
GUI says it's down to 12% or so.

Change-Id: I6924070419dda1ff0fc16e43404c37db1a38f9c9
Signed-off-by: Bill Richardson <wfrichar@chromium.org>