stm32l: add PLL clock configuration
Allow to run with the system clock connected to the PLL and the PLL
connected to the HSE (with a 16Mhz oscillator).
In that mode, the CPU clock is 32Mhz, and the USB peripheral is properly
clocked at 48Mhz.

But it keeps the HSI on in PLL mode :
The ADCs are using directly the HSI clock,
so we need to keep it on even though the system is clocked by the PLL
connected to the HSE.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=none
TEST=run on Reston with the PLL clock setting.

Change-Id: Id579687e1033e7f51c1013e6a8f3c71f931d2486