Update patch set 1
Patch Set 1:

> We don't have any ARMv7-M chips (that is, we never do -mcpu=arm7m).  All we have are Cortex-M,

That part is totally wrong.
ARMv7-M is the architecture and Cortex-M3 is the core implementation.
Cortex-M0 / Cortex-M3 / Cortex-M4 are all ARMv7-M architecture.

>  the instruction compiles fine there with the current chroot toolchain.

Yes the chroot toolchain has been updated.
A lot of external toolchains still don't support it.

> arm-none-eabi-gcc -mcpu=arm7m -mthumb -c foo2.c
>  arm-none-eabi-gcc -mcpu=arm7m -c foo2.c
>  armv7a-cros-linux-gnueabi-gcc -mcpu=arm7m -mthumb -c foo2.c

That command-line is wrong.
ARMv7-M is an architecture, the flag is:
-march=armv7-m

> so that's not a failure to encode SVC/SWI, it's just an unsupported combination.
> Even if we had a ARMv7-M core, it should still be fine,
>  because task.c is in core/cortex-m/, not core/armv7-m/

I disagree, we have ARMv7-M cores.

> I admit I may be missing a use case, but if so,
> please help me understand it better so I can refine the comment in task.c to avoid future confusion.

ok the actual comment is :
"use hardcoded opcode since SWI is not compiled properly on some older toolchains"

Patch-set: 1