Update cross-reference Makefile targets
Any of the tags, TAGS, or cscope.files targets will generate all three
cross-reference tables for the specified board. These are simply helpful for
source code browsing; the compiled images are unaffected.

BUG=chrome-os-partner:18343
BRANCH=none
TEST=manual

First, in the chroot:

  cd src/platform/ec
  make BOARD=link

Then, outside the chroot:

  make BOARD=link tags
  ls -ltr tags TAGS cscope.files

If you install the 'ctags' and 'etags' programs in the chroot, you can run
this command inside as well.

Change-Id: I8e16ef19b0f4e79aba887c308e58982fef8fa21f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>