Update patch set 2
Patch Set 2: Verified+1

I've modified the comment to include the workaround for older toolchains, in the unlikely event someone outside of Chromium needs to compile the code with one of those toolchains.

I think it's still correct to use SVC directly, since the toolchain in the chroot *does* support it.  It's cleaner and leads to correct disassembly.

Patch-set: 2
Label: Verified=+1