Add macro to pad sysjump data structs correctly
The data saved/restored using the system jump tags must be padded to a
multiple of four bytes. Currently this is done by hand, which makes
adjusting structs based on symbolic constants both difficult and
error-prone.

For example, this:

  static struct usb_state {
          uint8_t en[USB_PORT_COUNT];
          uint8_t pad[2]; /* Pad to 4 bytes for system_add_jump_tag(). */
  } state;

If we want to change USB_PORT_COUNT, we have to manually fix the padding
(and this particular example uses BUILD_ASSERT() to remind us).

This change replaces that kind of thing with this:

  DECLARE_JUMP_DATA(struct usb_state, {
          uint8_t en[USB_PORT_COUNT];
  });
  static struct usb_state state;

The padding is inserted silently and automatically, and doesn't require any
special attention if USB_PORT_COUNT is changed.

BUG=chrome-os-partner:18343
BRANCH=none
TEST=manual

To test, I added a global instance of each jump data struct, like this:

  struct usb_state HEY_usb_dumb;

I then compiled every platform, and looked at the size of each special
symbol. I did this before and after this change. The structure sizes were
unchanged:

  00000004 HEY_fan
  00000004 HEY_kbd
  00000004 HEY_kblight
  00000004 HEY_usb_dumb
  00000004 HEY_usb_smart

Change-Id: Iceab983eecb80472d0a1cc98933881c03f1b4d9b
Signed-off-by: Bill Richardson <wfrichar@chromium.org>