Update patch set 2
Patch Set 2:

Do the tags really need to be padded to 4-byte boundary ?
or the start of each tag needs to be aligned on the 4-byte boundary ?

If the constraint is the latter, a simple __attribute__((aligned(4))) would do and would be simpler.
if that's the former, you have a nice implementation for it.

Patch-set: 2
Reviewer: Gerrit User 1000016 <1000016@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0