Cleanly force battery to mAh mode
Smart batteries can report capacity in mAh or 10mW units.  We forced
the units to mAh in charge_state.c's main loop, but that doesn't
guarantee that they're actually set before the capacity is read.  It's
cleaner to check the capacity reporting mode when actually reading the
capacity.

BUG=chrome-os-partner:20881
BRANCH=none
TEST=battery command reports the same capacity data before/after change
     (on rambi, design=2940 mAh)

Change-Id: If22d2bc74ee71c7cecaa9ea41ffb66a42f4fe9ba
Signed-off-by: Randall Spangler <rspangler@chromium.org>