Update patch set 1
Patch Set 1: Code-Review-1

Better fix is to increase the size of the keyboard stack on Spring.

Prior to this change (and after the proposed revert), typing

ksstate on

would cause Spring to crash.  So reverting the console channel change merely masks the bug; it doesn't fix it.

Patch-set: 1
Label: Code-Review=-1