spring: Increase stack for keyscan task
The keyscan task is short enough on stack space that turning on
keyboard debugging causes a stack overflow.  This was previously not
the default, but https://chromium-review.googlesource.com/#/c/174373/
made it the default.  Reverting that change just masks the problem;
enabling keyboard debugging on Spring would still cause a stack
overflow.

Rather than reverting that change, increase the stack size of the
keyscan task so that it doesn't overflow.  There is sufficient space
to do this.  Even after increasing the keyboard stack from 256 bytes to
320 bytes and doing a 'sysjump rw' to force jump tags to populate,
'shmem' reports 132 bytes free.

BUG=chrome-os-partner:23834
BRANCH=none
TEST=Boot Spring.
     ksstate on
     Bang on keyboard for a bit
     taskinfo -> shows KEYSCAN task at 292/320 bytes free
     sysjump rw
     shmem -> shows 132 bytes free, 0 used

Change-Id: Idf9fdce5b9e6ca4d05d80a62ae9ea831ed508e3a
Signed-off-by: Randall Spangler <rspangler@chromium.org>