Update patch set 1
Patch Set 1:

(4 comments)

A few nits.  Looks good enough for a first pass after those are fixed.

Also, I note we're not very consistent in our indentation for python files in this project.  We should pick one indentation and stick with it.

For signing keys, can we just check in a dev/dummy signing key now?  Otherwise there's no way to use this script.  Or is there only one hard-coded private key?

Patch-set: 1