ite: Workaround that has UART working with interrupts
This is a workaround to use the UART with interrupts. This uses
the UART pins as GPIO to trigger an interrupt on the rising edge
of either the TX or RX pin, and then use this interrupt to check
if the UART module needs servicing.

Note: I've commented out starting up the console task when data
is received in uart_process_input().

DO NOT COMMIT.

Change-Id: I066093f037f2052924ce2fedd2fdfff2e6efd8e3
Signed-off-by: Alec Berg <alecaberg@chromium.org>