keyboard: Improve kbpress reliability for automation
The "kbpress" command had a few issues if you wanted to reliably use
it for automation.  Specifically it was not possible to guarantee how
much time would pass between the press of a key and the release of a
key.  Sometimes you might press and release before the key was
officially "there" and sometimes you might get a press and hold of a
key.

Fix this:
1. Make it so that kbpress with no press/release parameter gives a
   press and release (and guarantees that the press / release will
   actually take effect).
2. Make it so that kbpress guarantees that when it finishes that the
   key has actually been pressed or released.

BRANCH=pit
BUG=chrome-os-partner:24249
TEST=kbtype is (https://chromium-review.googlesource.com/178680) reliable
TEST=Pick Ibe00a796bde7d06416889b621359671a2f68e162 and test.

Change-Id: Ia213ab2e8d8da273e3ac4876d97d5452df88f47d
Signed-off-by: Doug Anderson <dianders@chromium.org>