Update patch set 1
Patch Set 1: Code-Review+2

LGTM.

The only minor concern to me is the usleep(debounce_up_us), which is 30ms. However, the fastest typematic repeat speed is 30HZ (33ms). This implies we could get 2+ keys when this repeat speed is set. But this is minor because I haven't never heard people set this so fast, and for Chrome OS the typematic is repeated by X rather than EC.

Patch-set: 1
Label: Code-Review=+2