Update patch set 1
Patch Set 1:

I was inconsistent with branch= tags on the recent on-branch patches for this bug. Thinking about it, I think branch=none is appropriate for an on-branch patch that wasn't cherry-picked over from ToT or another branch. In this case BRANCH=none indicates that this patch doesn't need to go anywhere else. In the future we may have patches landing on branches first that need to go up to the reference branch or ToT. In that case we could have something like BRANCH=slippy,master.

Patch-set: 1