Update patch set 1
Patch Set 1:

Louis and I talked.  It's unclear when we need to start counting that 200ms.  From power applied to PMIC, or deassertion of reset signal to PMIC.  So don't know if we need this delay on every power on, or just the first one.  And if power_on() is called after the EC has already been booted for 50ms, do we still need 200ms delay, or is 150ms (200-50) now sufficient?

Patch-set: 1