Update patch set 1
Patch Set 1:

This does change the compiled binary quite a bit on the branch.  It also seems to shave off about 64 bytes (in about a 55K image).

I'm not convinced that we want to introduce such a drastic change in an otherwise stable firmware, but we can see what others think.  In general the firmware on this branch is built with a builder that's using the old toolchain.

Patch-set: 1