clapper: Accelerometer driver for kxcj9.
WIP: Added driver for accelerometer, and a task to sample the
accelerometers.

Eventually, we will want to have another file at a higher abstraction
layer which calls into the accel driver to get accels and calculates
lid angle. But, for now, I have instrumented this code to give us an
idea of how the timing will work with reading the accelerometers.
This code has an EC console command 'accel' which can be used to change
sample rate and to display acceleration data to the console.

BUG=none
BRANCH=none
TEST=Ran on a clapper board. Can use the accel EC command to print out
accelerations to EC console.

Change-Id: Ieeaa4573b61d17327a778719a29da4dfd8fea09d
Signed-off-by: Alec Berg <alecaberg@chromium.org>