clapper: Accelerometer driver for kxcj9.
WIP: Added driver for accelerometer.

Also created a common file, motion_sense.c, to contain a task that
samples the accelerometers and calculates the lid angle. This file
is instrumented with calibration functions which will are conditionally
compiled in with CONFIG_ACCEL_CALIBRATE.

BUG=none
BRANCH=none
TEST=Ran on a clapper board. Can use the accel EC command to print out
accelerations to EC console. Built a mockup of a laptop by taping
Clapper board insid a box and taping acclerometers in place. Using
this setup, I verified the lid angle calibrations and calculations.

Change-Id: Ieeaa4573b61d17327a778719a29da4dfd8fea09d
Signed-off-by: Alec Berg <alecaberg@chromium.org>