Update patch set 1
Patch Set 1:

Branching strategy is that we keep the primary reference board in TOT and make sub-projects from a branch.  That allows partner eng to own the sub-projects.

I suppose we could leave it around forever, the way we have both falco and peppy in TOT, but it adds to the overhead of maintaining TOT.

In any case, create now so we can make progress, talk about deleting later when we get tired of too many boards in TOT again.

Patch-set: 1