thermal: dptf: Don't clear threshold condition on set
The DPTF framework will sometimes set thresholds and not expect
to get another event if the current temperature is above both the
previous threshold and the new threshold.

When a threshold is set only initialize the over condition if the
threshold was previously disabled in order to prevent it from
firing again.

BUG=None
Original-BUG=chrome-os-partner:23970
BRANCH=rambi
TEST=build and boot on rambi, start DPTF framework and observe
that when a threshold is crossed (going high) and a new threshold
is set at a lower value that it does not immediately result in a
new event.

Original-Change-Id: I6bad956fb5a49027a5c5f9c49a6fd68a0a96d693
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182004
(cherry picked from commit 24b39823b5bb3e207e1060fc684a3dec0f8d91c1)
Reviewed-on: https://chromium-review.googlesource.com/182077

Change-Id: Ic505a95500d58608ed5d99a15a2918a9cdde8963