Update patch set 5
Patch Set 5: Code-Review-1

If the GPIO simply drives a LED, then the signal name should indicate that.

We sometimes need to use LEDs for temporary debugging.  A GPIO like "POWER_LED_L" indicates this is ok, but "CHG_COMP" may mean the signal is connected to some other part of the charging system and could not be used for temporary debugging.

How about calling it CHARGING_LED_L?

Patch-set: 5
Label: Code-Review=-1