Update patch set 1
Patch Set 1:

(10 comments)

To dparker's question: I see there is a firmware-spring-3833.B branch after the series of 3824 spring branches. I take it that the 3833 branch didn't actually ship?

No - apparently 3833 was accidentally created - bhthompson confirmed that 3824 is the right branch (in addition to the current spring shipping firmware being 3824.129 which is what the skate fw branch was branched from)

Patch-set: 1
Label: Code-Review=0